refactor(ui-responsive): add common signature for the query match listener methods
Closes: INSTUI-3077
